Some initial updates
I'm adding and changing this as we go along and find more problems with this standard. This is a placehoder, of course, until we can get a real UX/UI designer to help us.
The ideas here are built with a UX designer from my place of work, but he just advised me generally and didn't actually write the standard or even looked at it.